feat: cluster and bugfixes
